Optimize unoptimized Index and Field nodes
When an Index or Field node is by itself Field will be faster if it doesn't have to check for the value being an Array. However, in runs of Index and Field it's probably faster to run them as a loop instead of as separate nodes, so the chaining combines them.