Switch links over to jmespath.org
The readthedocs links are for the python implementation of
jmespath, and now that there's a jmespath.org, I'm eventually
going to remove the readthedocslink for the specification and just
link to jmespath.org/specification.html.  There's already a
deprecation warning on the readthedocs link.

I also added a link to the jmespath tutorial. The previous link
to JMESPath examples was linking to the first examples in the spec,
which was just the examples for the first section, identifiers.