Restore support for legacy unquoted string literals
While we definitely want to transition away from unquoted string
literals, it seems reasonable to do that over time, similarly to how it
was handled in jmespath.py.

The reason for doing so, is to allow for a simple upgrade path, where a
new version of jmespath.rb can be installed, and the software using it
can transition all its configuration to use the new format, and then the
jmespath.rb version could be upgraded to not even support these
expressions.

The tests were copied verbatim from jmespath.py, and are not included in
the compliance test suite, so comes with an explicit license, which the
compliance test suite does not. JSON doesn't allow for comments, so I'm
not sure how to handle the attribution clause of the jmespath.py license