Removed simplecov usage.
Simplecov has a direct dependency on the `json` gem which makes
it impossible to test without the json gem present. It also forces
a newer version (>= 1.8) which also breaks the 1.9.3 tests with
and older json version.

We can explore having a separate run of the tests to generate
coverage, but for now, I'm removing simplecov.