Compare and determine the type of objects more efficiently
RubyObject has a #getNativeTypeIndex which is a much, much faster way of determining which type an object is, so use that instead of instanceof.

Also replace the generic implementation of #compare with something that uses RubyObject#compareTo for non-arrays and non-objects (for those to work jmespath-java probably has to be updated, because it expects -1 when these are non-equal).